# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API DOCUMENTATION",
        "title": "API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "prastiyo.beka12@gmail.com"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/api/v1/menu": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Menu Create",
                "parameters": [
                    {
                        "description": "Menu Create",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Menu.MenuRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/menu/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Menu Delete",
                "parameters": [
                    {
                        "description": "Menu Update",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Menu.MenuUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/menu/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Menu By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Menu By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Menu Update",
                "parameters": [
                    {
                        "description": "Menu Update",
                        "name": "menu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Menu.MenuUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/pemilu": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Pemilu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Pemilu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu Create",
                "parameters": [
                    {
                        "description": "Pemilu Create",
                        "name": "pemilu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Pemilu.PemiluRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/profile": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Profile Create",
                "parameters": [
                    {
                        "description": "Profile Create",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profile.ProfileRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/profile/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Profile Delete",
                "parameters": [
                    {
                        "description": "Profile Update",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profile.ProfileUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/profile/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Profile By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Profile By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Profile Update",
                "parameters": [
                    {
                        "description": "Profile Update",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profile.ProfileUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "Menu.MenuRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "translate": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "Menu.MenuUpdateRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "translate": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "Pemilu.PemiluRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info1": {
                    "type": "string"
                },
                "info2": {
                    "type": "string"
                },
                "info3": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "path_photo": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "total_voters": {
                    "type": "integer"
                },
                "total_voters_company": {
                    "type": "integer"
                },
                "total_voters_group": {
                    "type": "integer"
                },
                "total_voters_other": {
                    "type": "integer"
                },
                "tps_id": {
                    "type": "string"
                }
            }
        },
...
        "Profile.ProfileRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "city_id": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "district_id": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identity_number": {
                    "type": "string"
                },
                "info1": {
                    "type": "string"
                },
                "info2": {
                    "type": "string"
                },
                "info3": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path_photo": {
                    "type": "string"
                },
                "province_id": {
                    "type": "string"
                },
                "rt": {
                    "type": "string"
                },
                "rw": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "user_type_id": {
                    "type": "string"
                },
                "village_id": {
                    "type": "string"
                }
            }
        },
        "Profile.ProfileUpdateRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "city_id": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "district_id": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "identity_number": {
                    "type": "string"
                },
                "info1": {
                    "type": "string"
                },
                "info2": {
                    "type": "string"
                },
                "info3": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path_photo": {
                    "type": "string"
                },
                "province_id": {
                    "type": "string"
                },
                "rt": {
                    "type": "string"
                },
                "rw": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_type_id": {
                    "type": "string"
                },
                "village_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}